pyproject.toml
[tool.poetry]
name = "jpipe-runner"
version = "3.1.0"
description = "A Justification Runner designed for jPipe."
authors = [
    "Jason Lyu <xjasonlyu@gmail.com>",
    "Sébastien Mosser <mossers@mcmaster.ca>",
    "Baptiste Lacroix <contact@baptiste-lacroix.fr>"
]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/jpipe-mcscert/jpipe-runner"
keywords = ["justification", "pipeline", "ML", "AI"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
packages = [
    { include = "jpipe_runner", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.11"
networkx = "3.5"
pygraphviz = "1.14"
pyyaml = "6.0.2"

matplotlib = { version = "3.10.5", optional = true }
sphinx = { version = "8.1.3", optional = true }

[tool.poetry.extras]
gui = ["matplotlib"]
docs = ["sphinx"]
full = ["matplotlib", "sphinx"]

[tool.poetry.scripts]
jpipe-runner = "jpipe_runner.runner:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-mock = "^3.14.1"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
